+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Unreleased
 - Add msg app for LCD/Windows notifications
 - Remove environment sigils integration
 - Show full date and timezone in admin clock tooltip
+- Fallback to site domain in admin badge when display name missing
 
 0.1.1 [revision 76f70b6a72c78fcdf143a19ddcc88a0fbd209b3d]
 ---------------------------------------------------------